TTALE - 189 : Added new profile interface and cleaned up other API and classes
