+++ src/main/java/org/jboss/tattletale/analyzers/JarScanner.java
@@ -24,6 +24,7 @@ package org.jboss.tattletale.analyzers;
 import org.jboss.tattletale.core.Archive;
 import org.jboss.tattletale.core.JarArchive;
 import org.jboss.tattletale.core.Location;
+import org.jboss.tattletale.profiles.Profile;
 
 import java.io.File;
 import java.io.IOException;
@@ -72,7 +73,7 @@ public class JarScanner extends AbstractScanner
     * @param blacklisted The set of black listed packages
     * @return The archive
     */
-   public Archive scan(File file, Map<String, SortedSet<String>> gProvides, List<Archive> known,
+   public Archive scan(File file, Map<String, SortedSet<String>> gProvides, List<Profile> known,
                        Set<String> blacklisted)
    {
       Archive archive = null;
