+++ src/main/java/org/jboss/tattletale/profiles/JavaEE6.java
@@ -19,36 +19,30 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.tattletale.reporting.profiles;
+package org.jboss.tattletale.profiles;
 
 import org.jboss.tattletale.core.ArchiveTypes;
-
-import java.util.Set;
 
 import javassist.bytecode.ClassFile;
 
 /**
- * Sun: Java 6
+ * Java Enterprise 6
  *
  * @author Jesper Pedersen <jesper.pedersen@jboss.org>
  */
-public class SunJava6 extends CommonProfile
+public class JavaEE6 extends AbstractProfile
 {
-
-   private static final String CLASS_SET = "sunjdk6.clz.gz";
-   private static final String PROFILE_NAME = "Sun Java 6";
-   private static final String PROFILE_CODE = "java6";
-   private static final String PROFILE_LOCATION = "rt.jar";
+   private static final String CLASS_SET = "ee6.clz.gz";
+   private static final String PROFILE_NAME = "Java Enterprise 6";
+   private static final String PROFILE_CODE = "ee6";
+   private static final String PROFILE_LOCATION = "javaee-api-6.jar";
    private static final int ARCHIVE_TYPE = ArchiveTypes.JAR;
    private static final int CLASSFILE_VERSION = ClassFile.JAVA_6;
 
    /** Constructor */
-   public SunJava6()
+   public JavaEE6()
    {
       super(CLASS_SET, ARCHIVE_TYPE, PROFILE_NAME, CLASSFILE_VERSION, PROFILE_LOCATION);
-
-      addSubArchive(new SunJava6JCE());
-      addSubArchive(new SunJava6JSSE());
    }
 
    @Override
@@ -61,12 +55,5 @@ public class SunJava6 extends CommonProfile
    protected String getProfileName()
    {
       return PROFILE_NAME;
-   }
-
-   @Override
-   public boolean included(boolean allProfiles, Set<String> profileSet)
-   {
-      return allProfiles || profileSet == null || (profileSet.contains(getProfileCode())
-            || profileSet.contains(getProfileName()));
    }
 }
