+++ src/main/java/org/jboss/tattletale/profiles/SunJava5JCE.java
@@ -19,29 +19,29 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.tattletale.reporting.profiles;
+package org.jboss.tattletale.profiles;
 
 import org.jboss.tattletale.core.ArchiveTypes;
 
 import javassist.bytecode.ClassFile;
 
 /**
- * Sun: Java 5 (JSSE)
+ * Sun: Java 5 (JCE)
  *
  * @author Jesper Pedersen <jesper.pedersen@jboss.org>
  */
-public class SunJava6JSSE extends CommonProfile
+public class SunJava5JCE extends AbstractProfile
 {
 
-   private static final String CLASS_SET = "sunjdk6-jsse.clz.gz";
-   private static final String PROFILE_NAME = "Sun Java 6 (JSSE)";
-   private static final String PROFILE_CODE = "jsse5";
-   private static final String PROFILE_LOCATION = "jsse.jar";
+   private static final String CLASS_SET = "sunjdk5-jce.clz.gz";
+   private static final String PROFILE_NAME = "Sun Java 5 (JCE)";
+   private static final String PROFILE_CODE = "jce5";
+   private static final String PROFILE_LOCATION = "jce.jar";
    private static final int ARCHIVE_TYPE = ArchiveTypes.JAR;
-   private static final int CLASSFILE_VERSION = ClassFile.JAVA_6;
+   private static final int CLASSFILE_VERSION = ClassFile.JAVA_5;
 
    /** Constructor */
-   public SunJava6JSSE()
+   public SunJava5JCE()
    {
       super(CLASS_SET, ARCHIVE_TYPE, PROFILE_NAME, CLASSFILE_VERSION, PROFILE_LOCATION);
    }
