+++ src/main/java/org/jboss/tattletale/profiles/SunJava5JSSE.java
@@ -19,28 +19,29 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.tattletale.reporting.profiles;
+package org.jboss.tattletale.profiles;
 
 import org.jboss.tattletale.core.ArchiveTypes;
 
 import javassist.bytecode.ClassFile;
 
 /**
- * Java Enterprise 6
+ * Sun: Java 5 (JSSE)
  *
  * @author Jesper Pedersen <jesper.pedersen@jboss.org>
  */
-public class JavaEE6 extends CommonProfile
+public class SunJava5JSSE extends AbstractProfile
 {
-   private static final String CLASS_SET = "ee6.clz.gz";
-   private static final String PROFILE_NAME = "Java Enterprise 6";
-   private static final String PROFILE_CODE = "ee6";
-   private static final String PROFILE_LOCATION = "javaee-api-6.jar";
+
+   private static final String CLASS_SET = "sunjdk5-jsse.clz.gz";
+   private static final String PROFILE_NAME = "Sun Java 5 (JSSE)";
+   private static final String PROFILE_CODE = "jsse5";
+   private static final String PROFILE_LOCATION = "jsse.jar";
    private static final int ARCHIVE_TYPE = ArchiveTypes.JAR;
-   private static final int CLASSFILE_VERSION = ClassFile.JAVA_6;
+   private static final int CLASSFILE_VERSION = ClassFile.JAVA_5;
 
    /** Constructor */
-   public JavaEE6()
+   public SunJava5JSSE()
    {
       super(CLASS_SET, ARCHIVE_TYPE, PROFILE_NAME, CLASSFILE_VERSION, PROFILE_LOCATION);
    }
